+++ src/test/java/com/google/security/zynamics/reil/translators/x86/RepStosdTranslatorTest.java
@@ -76,7 +76,7 @@ public class RepStosdTranslatorTest {
     assertEquals(BigInteger.valueOf(0x1004), interpreter.getVariableValue("edi"));
 
     assertEquals(BigInteger.valueOf(4L), BigInteger.valueOf(interpreter.getMemorySize()));
-    assertEquals(0x12345678, interpreter.getMemory().load(0x1000, 4));
+    assertEquals(BigInteger.valueOf(0x12345678), interpreter.getMemory().load(0x1000, 4));
   }
 
   @Test
@@ -103,6 +103,6 @@ public class RepStosdTranslatorTest {
     assertEquals(BigInteger.valueOf(0x0FFC), interpreter.getVariableValue("edi"));
 
     assertEquals(BigInteger.valueOf(4L), BigInteger.valueOf(interpreter.getMemorySize()));
-    assertEquals(0x12345678, interpreter.getMemory().load(0x1000, 4));
+    assertEquals(BigInteger.valueOf(0x12345678), interpreter.getMemory().load(0x1000, 4));
   }
 }
